pyproject.toml
[tool.poetry]
name = "e-commerce-using-fastapi-vue-js"
version = "0.1.0"
description = ""
authors = ["Isaiah-Mwinga <motongori@marketforce360.com>"]
readme = "README.md"
packages = [{include = "e_commerce_using_fastapi_vue_js"}]

[tool.poetry.dependencies]
python = "^3.10"
pendulum = "^2.1"
alembic = "^1.8.1"
pip = "^22.3.1"
install = "^1.3.5"
FastAPI-SQLAlchemy = "^0.2.1"
pytest = "^7.2.0"
dotenv = "^0.0.5"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
fastapi-jwt-auth = "^0.5.0"
email-validator = "^1.3.0"
passlib = "^1.7.4"
dotenv-python = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
